+++ src/Faqt/HttpResponseMessageAssertions.fs
@@ -7,6 +7,7 @@ open System.Net.Http
 open System.Runtime.CompilerServices
 open System.Text
 open Faqt.AssertionHelpers
+open Faqt.Configuration
 
 
 module private HttpHeaderValueParsing =
@@ -545,7 +546,7 @@ type HttpResponseMessageAssertions =
         let fail () =
             t
                 .With("Header", name)
-                .With("Value", value)
+                .With("Value", Config.Current.MapHttpHeaderValues name value)
                 .With("Response", t.Subject)
                 .With("Request", t.Subject.RequestMessage)
                 .Fail(because)
@@ -573,14 +574,14 @@ type HttpResponseMessageAssertions =
             | Some value ->
                 t
                     .With("Header", name)
-                    .With("But was present with value", value)
+                    .With("But was present with value", Config.Current.MapHttpHeaderValues name value)
                     .With("Response", t.Subject)
                     .With("Request", t.Subject.RequestMessage)
                     .Fail(because)
             | None ->
                 t
                     .With("Header", name)
-                    .With("But was present with values", values)
+                    .With("But was present with values", values |> List.map (Config.Current.MapHttpHeaderValues name))
                     .With("Response", t.Subject)
                     .With("Request", t.Subject.RequestMessage)
                     .Fail(because)